LUM-828: Internal user credit endpoint (#31)
* LUM-828: Internal user credit endpoint
* LUM-841: Automate deployments (#33)

+++ src/app/schemas/usage.py
@@ -1,5 +1,4 @@
 from datetime import date
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, Field, ConfigDict
@@ -15,9 +14,9 @@ class UsageRecordResponse(BaseModel):
     id: UUID = Field(..., description="The unique identifier of the usage record")
     created_at: DateTime = Field(..., description="The timestamp when the usage record was created")
     service_name: ServiceName = Field(..., description="The name of the service used")
-    usage_amount: Decimal = Field(..., description="The amount of usage for the service")
+    usage_amount: float = Field(..., description="The amount of usage for the service")
     usage_unit: UsageUnit = Field(..., description="The unit of usage for the service")
-    cost: Decimal = Field(..., description="The cost associated with the usage")
+    cost: float = Field(..., description="The cost associated with the usage")
     fine_tuning_job_name: str = Field(..., description="The name of the associated fine-tuning job")
     model_config = ConfigDict(from_attributes=True)
 
@@ -28,5 +27,5 @@ class TotalCostResponse(BaseModel):
     """
     start_date: date = Field(..., description="The start date of the period for which the cost is calculated")
     end_date: date = Field(..., description="The end date of the period for which the cost is calculated")
-    total_cost: Decimal = Field(..., description="The total cost for the specified period")
+    total_cost: float = Field(..., description="The total cost for the specified period")
     model_config = ConfigDict(from_attributes=True)
