+++ src/app/schemas/user.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, EmailStr, Field, ConfigDict
@@ -24,5 +23,5 @@ class UserResponse(BaseModel):
     status: UserStatus = Field(..., description="The current status of the user")
     name: str = Field(..., description="The name of the user")
     email: EmailStr = Field(..., description="The email address of the user")
-    credits_balance: Decimal = Field(..., description="The current credit balance of the user")
+    credits_balance: float = Field(..., description="The current credit balance of the user")
     model_config = ConfigDict(from_attributes=True)
